+++ helpers/NPWindows.js
@@ -299,10 +299,11 @@ export async function openNoteInNewSplitIfNeeded(filename: string): Promise<bool
 /**
  * Open a calendar note in a split editor, and (optionally) move insertion point to 'cursorPointIn'
  * @author @jgclark
- * @param {*} filename
- * @param {*} cursorPointIn
+ * @param {string} filename
+ * @param {string | number} cursorPointIn
  */
 export async function openCalendarNoteInSplit(filename: string, cursorPointIn?: string | number = 0): Promise<void> {
+  logDebug('openCalendarNoteInSplit', `Opening calendar note '${filename}' in split at cursor point ${cursorPointIn}`)
   // For some reason need to add a bit to get to the right place.
   const cursorPoint = (typeof cursorPointIn === 'string') ? parseInt(cursorPointIn) + 21 : cursorPointIn + 21
   const res = Editor.openNoteByDateString(filename.split('.')[0], false, cursorPoint, cursorPoint, true)